+++ packages/golden-layout/src/js/controls/Header.js
@@ -133,10 +133,12 @@ lm.utils.copy(lm.controls.Header.prototype, {
 		}
 
 		this.tabs.splice(index, 0, tab);
+		this._updateTabSizesWithoutDropdown();
 	},
 	moveTab: function (fromIndex, toIndex) {
 		lm.utils.moveInArray(this.tabs, fromIndex, toIndex);
 
+		this._updateTabSizesWithoutDropdown();
 		this.parent._syncContentItemOrder();
 	},
 
@@ -484,6 +486,59 @@ lm.utils.copy(lm.controls.Header.prototype, {
 				tabElement.css({ 'z-index': 'auto', 'margin-left': '' });
 				this.tabsContainer.append(tabElement);
 			}
+		}
+
+	},
+
+	/**
+	 * Applies the necessary tab styles without using a dropdown
+	 *
+	 * @returns {void}
+	 */
+	 _updateTabSizesWithoutDropdown: function (showTabMenu) {
+		if (this.tabs.length === 0) {
+			return;
+		}
+
+		var size = function (val) {
+			return val ? 'width' : 'height';
+		};
+		this.element.css(size(!this.parent._sided), '');
+		// this.element[size(this.parent._sided)](this.layoutManager.config.dimensions.headerHeight);
+		var availableWidth = this.element.outerWidth() - this.controlsContainer.outerWidth() - this._tabControlOffset,
+			cumulativeTabWidth = 0,
+			visibleTabWidth = 0,
+			tabElement,
+			i,
+			j,
+			marginLeft,
+			overlap = 0,
+			tabWidth,
+			tabOverlapAllowance = this.layoutManager.config.settings.tabOverlapAllowance,
+			tabOverlapAllowanceExceeded = false,
+			activeIndex = (this.activeContentItem ? this.tabs.indexOf(this.activeContentItem.tab) : 0),
+			activeTab = this.tabs[activeIndex];
+		if (this.parent._sided)
+			availableWidth = this.element.outerHeight() - this.controlsContainer.outerHeight() - this._tabControlOffset;
+
+		for (i = 0; i < this.tabs.length; i++) {
+			tabElement = this.tabs[i].element;
+
+			//Put the tab in the tabContainer so its true width can be checked
+			this.tabsContainer.append(tabElement);
+			tabWidth = tabElement.outerWidth() + parseInt(tabElement.css('margin-right'), 10);
+
+			cumulativeTabWidth += tabWidth;
+
+			//Include the active tab's width if it isn't already
+			//This is to ensure there is room to show the active tab
+			if (activeIndex <= i) {
+				visibleTabWidth = cumulativeTabWidth;
+			} else {
+				visibleTabWidth = cumulativeTabWidth + activeTab.element.outerWidth() + parseInt(activeTab.element.css('margin-right'), 10);
+			}
+
+			tabElement.css({ 'z-index': 'auto', 'margin-left': '' });
 		}
 
 	},